backend/ml_server/requirements.txt
# Core ML libraries
numpy>=1.21.0,<2.0.0
torch>=2.0.0
torchvision>=0.15.0

# OCR libraries
easyocr>=1.7.1
pytesseract>=0.3.10

# Image processing and augmentation
opencv-python>=4.8.1.78
Pillow>=10.0.1
albumentations==1.3.1
imgaug>=0.4.0

# Web framework
flask>=2.3.3
flask-cors>=4.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.1
pandas>=2.0.0
scikit-learn>=1.3.0

# Development tools
pytest>=7.4.0
black>=23.7.0
joblib>=1.3.1